+++ metadata/src/audio/item.rs
@@ -1,4 +1,4 @@
-use std::fmt::Debug;
+use std::{fmt::Debug, path::PathBuf};
 
 use crate::{
     Metadata,
@@ -50,6 +50,16 @@ pub enum UniqueFields {
         number: u32,
         disc_number: u32,
     },
+    Local {
+        // artists / album_artists can't be a Vec here, they are retrieved from metadata as a String,
+        // and we cannot make any assumptions about them being e.g. comma-separated
+        artists: Option<String>,
+        album: Option<String>,
+        album_artists: Option<String>,
+        number: Option<u32>,
+        disc_number: Option<u32>,
+        path: PathBuf,
+    },
     Episode {
         description: String,
         publish_time: Date,
@@ -77,7 +87,7 @@ impl AudioItem {
                     return Err(Error::unavailable(MetadataError::ExplicitContentFiltered));
                 }
 
-                let uri_string = uri.to_uri()?;
+                let uri_string = uri.to_uri();
                 let album = track.album.name;
 
                 let album_artists = track
@@ -146,7 +156,7 @@ impl AudioItem {
                     return Err(Error::unavailable(MetadataError::ExplicitContentFiltered));
                 }
 
-                let uri_string = uri.to_uri()?;
+                let uri_string = uri.to_uri();
 
                 let covers = get_covers(episode.covers, image_url);