+++ CHANGELOG.md
@@ -7,18 +7,46 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- [core] Made `SpotifyId::to_base62`, `SpotifyId::to_base16`, `FileId::to_base16`, `SpotifyUri::to_id`, `SpotifyUri::to_uri` infallible (breaking)
+
+### Fixed
+
+- [main] Fixed `--volume-ctrl fixed` not disabling volume control
+
+## [0.8.0] - 2025-11-10
+
 ### Added
 
+- [connect] Add method `transfer` to `Spirc` to automatically transfer the playback to ourselves
+- [core] Add method `transfer` to `SpClient`
 - [core] Add `SpotifyUri` type to represent more types of URI than `SpotifyId` can
+- [discovery] Add support for [device aliases](https://developer.spotify.com/documentation/commercial-hardware/implementation/guides/zeroconf#device-aliases)
+- [main] `--local-file-dir` / `-l` option added to binary to specify local file directories to pull from
+- [metadata] `Local` variant added to `UniqueFields` enum (breaking)
+- [playback] Local files can now be played with the following caveats:
+  - They must be sampled at 44,100 Hz
+  - They cannot be played from a Connect device using the dedicated 'Local Files' playlist; they must be added to another playlist first
+- [playback] `local_file_directories` field added to `PlayerConfig` struct (breaking)
 
 ### Changed
 
-- [playback] Changed type of `SpotifyId` fields in `PlayerEvent` members to `SpotifyUri` (breaking)
+- [contrib] Switched contrib/Dockerfile to new Debian stable (trixie)
+- [core] `get_radio_for_track` function changed from accepting a `SpotifyId` to accepting a `SpotifyUri` (breaking)
+- [core] Changed return type of `get_extended_metadata` to return `BatchedExtensionResponse` (breaking)
+- [core] Changed parameter of `get_<item>_metadata` from `SpotifyId` to `SpotifyUri` (breaking)
 - [metadata] Changed arguments for `Metadata` trait from `&SpotifyId` to `&SpotifyUri` (breaking)
-- [player] `load` function changed from accepting a `SpotifyId` to accepting a `SpotifyUri` (breaking)
-- [player] `preload` function changed from accepting a `SpotifyId` to accepting a `SpotifyUri` (breaking)
-- [spclient] `get_radio_for_track` function changed from accepting a `SpotifyId` to accepting a `SpotifyUri` (breaking)
+- [playback] Changed type of `SpotifyId` fields in `PlayerEvent` members to `SpotifyUri` (breaking)
+- [playback] `load` function changed from accepting a `SpotifyId` to accepting a `SpotifyUri` (breaking)
+- [playback] `preload` function changed from accepting a `SpotifyId` to accepting a `SpotifyUri` (breaking)
 
+### Fixed
+
+- [connect] Fixed failed transferring with transfer data that had an empty context uri and no tracks
+- [connect] Use the provided index or the first as fallback value to always play a track on loading
+- [core] Fixed a problem where the metadata didn't include the audio file by switching to `get_extended_metadata`
+- [core] Fixed connection issues after system suspend on Linux
 
 ### Removed
 
@@ -38,6 +66,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- [core] Fix issue where building with native-tls would fail
 - [connect] Repeat context will not go into autoplay anymore and triggering autoplay while shuffling shouldn't reshuffle anymore
 - [connect] Only deletes the connect state on dealer shutdown instead on disconnecting
 - [core] Fixed a problem where in `spclient` where an HTTP/411 error was thrown because the header was set wrong
@@ -158,7 +187,7 @@ will be well worth it.
 
 All these changes are likely to introduce new bugs as well as some regressions.
 We appreciate all your testing and contributions to the repository:
-https://github.com/librespot-org/librespot
+<https://github.com/librespot-org/librespot>
 
 ### Changed
 
@@ -441,7 +470,8 @@ v0.4.x as a stable branch until then.
 
 ## [0.1.0] - 2019-11-06
 
-[unreleased]: https://github.com/librespot-org/librespot/compare/v0.7.1...HEAD
+[unreleased]: https://github.com/librespot-org/librespot/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/librespot-org/librespot/compare/v0.7.1...v0.8.0
 [0.7.1]: https://github.com/librespot-org/librespot/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/librespot-org/librespot/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/librespot-org/librespot/compare/v0.5.0...v0.6.0